requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
effdet==0.4.1
emoji==2.14.1
faiss-cpu==1.11.0.post1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.8
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.47.1
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.0.0
pikepdf==9.10.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
pypdf==5.8.0
pypdfium2==4.30.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.17
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.9
unstructured-client==0.39.1
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
